Changing the structure of the project. Now the authentication functions are in a separate package to improve code re-usability
